Split lsp/server.py: extract commands, workspace_init, lifecycle modules
Phase 5-7 of the server.py decomposition:
- lsp/commands.py: all workspace/executeCommand handlers (optimise, minify,
  irule metadata, safe-fix helpers, package suggestions, BIG-IP extraction,
  compiler explorer, tclpkg, dialect switching, config export)
- lsp/workspace_init.py: background scanning, dialect auto-upgrade, progress
  reporting, on_initialized handler
- lsp/lifecycle.py: did_open/did_change/did_close, shutdown, file-watch,
  file rename hooks

server.py reduced from 4122 → ~1177 lines; all handlers now thin wrappers
calling into the extracted modules. Full test suite: 10,493 passed.
