+++ lsp/server.py
@@ -16,29 +16,7 @@ import core.common.codes_all  # noqa: F401  # trigger all code registrations
 from core.analysis import analyse
 from core.analysis.irules_checks import DEFAULT_GENERIC_VARIABLE_PATTERNS
 from core.commands.registry import REGISTRY
-from core.commands.registry.info import effective_event_requires
-from core.commands.registry.namespace_registry import NAMESPACE_REGISTRY as EVENT_REGISTRY
-from core.commands.registry.runtime import configure_signatures, is_irules_dialect
-from core.common.document_buffer import DocumentBuffer
 from core.common.lsp import to_lsp_location
-from core.common.optimisation_profiles import (
-    DEFAULT_ACTION_PROFILE,
-    DEFAULT_EDITOR_PROFILE,
-    PROFILE_NAMES,
-    profile_from_name,
-    profile_to_disabled,
-    resolve_profile,
-)
-from core.common.user_config import (
-    get_all_settings,
-    load_user_config,
-    save_settings_to_config,
-)
-from core.compiler.optimiser import optimise_source
-from core.formatting import FormatterConfig
-from core.minifier import minify_tcl
-from explorer.pipeline import run_pipeline as explorer_run_pipeline
-from explorer.serialise import serialise_result as explorer_serialise_result
 
 from .features import (
     SEMANTIC_TOKEN_MODIFIERS,
@@ -60,7 +38,7 @@ from .features.code_lens import get_code_lenses, resolve_code_lens
 from .features.completion import get_completions
 from .features.declaration import get_declaration
 from .features.definition import get_bigip_definition, get_definition
-from .features.diagnostics import get_basic_diagnostics, get_deep_diagnostics, get_diagnostics
+from .features.diagnostics import get_deep_diagnostics, get_diagnostics
 from .features.document_highlight import get_document_highlights
 from .features.document_links import get_document_links
 from .features.document_symbols import get_document_symbols
@@ -70,7 +48,6 @@ from .features.hover import get_hover
 from .features.implementation import get_implementations
 from .features.inlay_hints import get_inlay_hints
 from .features.linked_editing_range import get_linked_editing_ranges
-from .features.package_suggestions import rank_package_suggestions
 from .features.references import find_proc_call_sites, get_references
 from .features.rename import get_rename_edits, prepare_rename
 from .features.selection_range import get_selection_ranges
@@ -86,15 +63,12 @@ from .features.type_hierarchy import (
 from .features.type_hierarchy import (
     supertypes as get_supertypes,
 )
-from .features.workspace_file_ops import compute_batch_rename_edits
 from .features.workspace_symbols import get_workspace_symbols
-from .workspace.scanner import path_to_uri, uri_to_path
-from .workspace.workspace_index import EntrySource, WorkspaceIndex
+from .workspace.workspace_index import WorkspaceIndex
 
 log = logging.getLogger(__name__)
 
 
-from .feature_config import FeatureConfig
 
 
 try:
@@ -255,6 +229,9 @@ server.protocol._handle_notification = _log_notification  # type: ignore[assignm
 
 import lsp.state as _state
 import lsp.diagnostics_pipeline as _dp
+import lsp.commands as _commands
+import lsp.workspace_init as _workspace_init
+import lsp.lifecycle as _lifecycle
 from .state import (
     workspace_state,
     workspace_index,
@@ -277,6 +254,9 @@ from .state import (
 
 _state.configure(server)
 _dp.configure(server)
+_commands.register(server)
+_workspace_init.register(server)
+_lifecycle.register(server)
 
 
 # Capabilities
@@ -1111,691 +1091,6 @@ def on_code_action(
             action.edit.changes = remapped
     return actions or None
 
-
-# Workspace commands
-
-
-@server.command("tcl-lsp.optimiseDocument")
-def on_optimise_document(uri: str, profile: str = "full") -> dict | None:
-    from core.compiler.optimiser import optimise_source_multipass
-
-    source = _get_doc_source(uri)
-    disabled, multi_pass, max_iterations = resolve_profile(
-        profile,
-        default=DEFAULT_ACTION_PROFILE,
-    )
-    if multi_pass:
-        optimised, opts, _iters = optimise_source_multipass(
-            source,
-            max_iterations=max_iterations,
-            disabled=disabled,
-        )
-    else:
-        optimised, opts = optimise_source(source, disabled=disabled)
-    items = []
-    for o in opts:
-        item: dict = {
-            "code": o.code,
-            "message": o.message,
-            "startLine": o.range.start.line,
-            "startCharacter": o.range.start.character,
-            "endLine": o.range.end.line,
-            "endCharacter": o.range.end.character,
-            "replacement": o.replacement,
-        }
-        if o.group is not None:
-            item["group"] = o.group
-        if o.hint_only:
-            item["hintOnly"] = True
-        items.append(item)
-    return {
-        "optimisations": items,
-        "source": optimised,
-    }
-
-
-@server.command("tcl-lsp.minifyDocument")
-def on_minify_document(
-    uri: str, compact: bool = False, aggressive: bool = False, isolated: bool = False
-) -> dict | None:
-    """Minify the Tcl document: strip comments, collapse whitespace, join commands."""
-    source = _get_doc_source(uri)
-    if aggressive:
-        result = minify_tcl(source, aggressive=True, isolated=isolated)
-        return {
-            "source": result.source,
-            "originalLength": result.original_length,
-            "minifiedLength": result.minified_length,
-            "symbolMap": result.symbol_map.format(),
-            "optimisationsApplied": result.optimisations_applied,
-        }
-    if compact:
-        minified, symbol_map = minify_tcl(source, compact_names=True, isolated=isolated)
-        return {
-            "source": minified,
-            "originalLength": len(source),
-            "minifiedLength": len(minified),
-            "symbolMap": symbol_map.format(),
-        }
-    minified = minify_tcl(source)
-    return {
-        "source": minified,
-        "originalLength": len(source),
-        "minifiedLength": len(minified),
-    }
-
-
-@server.command("tcl-lsp.unminifyError")
-def on_unminify_error(
-    error_message: str,
-    symbol_map: str,
-    minified_source: str = "",
-    original_source: str = "",
-) -> dict:
-    """Translate a minified-code error message back to original names."""
-    from core.minifier import unminify_error
-
-    translated = unminify_error(
-        error_message,
-        symbol_map=symbol_map,
-        minified_source=minified_source or None,
-        original_source=original_source or None,
-    )
-    return {
-        "originalError": error_message,
-        "translatedError": translated,
-        "changed": translated != error_message,
-    }
-
-
-@server.command("tcl-lsp.describeIruleEvent")
-def on_describe_irule_event(event_name: str) -> dict:
-    """Return deterministic registry metadata for an iRules event."""
-    event = (event_name or "").strip()
-    when_values = REGISTRY.argument_values("when", 0, "f5-irules")
-    known_events = {value.value for value in when_values}
-    deprecated_events = {
-        value.value for value in when_values if "deprecated" in (value.detail or "").lower()
-    }
-
-    is_known = event in known_events
-    if not event or not is_known:
-        valid_commands: list[str] = []
-    else:
-        event_set = REGISTRY.commands_for_event("f5-irules", event)
-        valid_commands = sorted(event_set.valid_commands)
-    return {
-        "event": event,
-        "known": is_known,
-        "deprecated": event in deprecated_events,
-        "validCommandCount": len(valid_commands),
-        "sampleCommands": valid_commands[:80],
-    }
-
-
-@server.command("tcl-lsp.describeIruleCommand")
-def on_describe_irule_command(command_name: str) -> dict:
-    """Return deterministic registry metadata for an iRules command."""
-    name = (command_name or "").strip()
-    spec = REGISTRY.get(name, "f5-irules")
-    if spec is None and name:
-        lowered = name.lower()
-        for candidate in REGISTRY.command_names("f5-irules"):
-            if candidate.lower() == lowered:
-                spec = REGISTRY.get(candidate, "f5-irules")
-                name = candidate
-                break
-
-    if spec is None:
-        return {
-            "found": False,
-            "command": name,
-        }
-
-    synopsis = tuple(spec.hover.synopsis) if spec.hover and spec.hover.synopsis else ()
-
-    result: dict = {
-        "found": True,
-        "command": name,
-        "summary": spec.hover.summary if spec.hover else "",
-        "synopsis": list(synopsis),
-        "switches": list(spec.switch_names()),
-    }
-
-    requires = effective_event_requires(name, spec.event_requires, dialect="f5-irules")
-    if requires is not None:
-        result["validEvents"] = EVENT_REGISTRY.events_matching(requires)
-        result["anyEvent"] = not (
-            requires.client_side
-            or requires.server_side
-            or requires.transport is not None
-            or bool(requires.profiles)
-            or bool(requires.also_in)
-            or requires.init_only
-            or requires.flow
-            or requires.capability is not None
-        )
-        result["eventRequires"] = {
-            "clientSide": requires.client_side,
-            "serverSide": requires.server_side,
-            "transport": requires.transport,
-            "profiles": sorted(requires.profiles),
-        }
-
-    return result
-
-
-def _apply_non_overlapping_fixes(
-    source: str,
-    fixes: list[tuple[int, int, str, str, str]],
-) -> tuple[str, list[dict[str, str]]]:
-    """Apply non-overlapping text replacements using source offsets."""
-    valid = [fix for fix in fixes if 0 <= fix[0] <= fix[1] < len(source)]
-    valid.sort(key=lambda item: (item[0], item[1]))
-
-    accepted: list[tuple[int, int, str, str, str]] = []
-    current_end = -1
-    for fix in valid:
-        start, end, _text, _code, _description = fix
-        if start <= current_end:
-            continue
-        accepted.append(fix)
-        current_end = end
-
-    rewritten = source
-    applied: list[dict[str, str]] = []
-    for start, end, new_text, code, description in reversed(accepted):
-        rewritten = rewritten[:start] + new_text + rewritten[end + 1 :]
-        applied.append(
-            {
-                "code": code,
-                "description": description or "Apply safe fix",
-            }
-        )
-    applied.reverse()
-    return rewritten, applied
-
-
-def _collect_pending_safe_fixes(
-    source: str,
-) -> list[tuple[int, int, str, str, str]]:
-    """Collect first safe fix per diagnostic for *source*."""
-    analysis = analyse(source)
-    pending: list[tuple[int, int, str, str, str]] = []
-    for diagnostic in analysis.diagnostics:
-        if diagnostic.code not in _SAFE_FIX_CODES:
-            continue
-        if not diagnostic.fixes:
-            continue
-        fix = diagnostic.fixes[0]
-        pending.append(
-            (
-                fix.range.start.offset,
-                fix.range.end.offset,
-                fix.new_text,
-                diagnostic.code,
-                fix.description,
-            )
-        )
-    return pending
-
-
-def _apply_safe_fixes_iteratively(
-    source: str,
-    *,
-    max_passes: int = 4,
-) -> tuple[str, list[dict[str, str]]]:
-    """Apply safe fixes in multiple passes until source stabilises."""
-    rewritten = source
-    applied_total: list[dict[str, str]] = []
-    for _ in range(max_passes):
-        pending = _collect_pending_safe_fixes(rewritten)
-        if not pending:
-            break
-        next_source, applied = _apply_non_overlapping_fixes(rewritten, pending)
-        if not applied or next_source == rewritten:
-            break
-        applied_total.extend(applied)
-        rewritten = next_source
-    return rewritten, applied_total
-
-
-@server.command("tcl-lsp.fixAllSafeIssues")
-def on_fix_all_safe_issues(uri: str) -> dict | None:
-    """Apply all non-overlapping safe code fixes for a document."""
-    source = _get_doc_source(uri)
-    rewritten, applied = _apply_safe_fixes_iteratively(source)
-    return {
-        "source": rewritten,
-        "applied": applied,
-    }
-
-
-@server.command("tcl-lsp.listIruleEvents")
-def on_list_irule_events() -> dict:
-    """Return all known iRules event names from the registry."""
-    when_values = REGISTRY.argument_values("when", 0, "f5-irules")
-    events = sorted(v.value for v in when_values if v.value)
-    return {
-        "events": events,
-        "count": len(events),
-    }
-
-
-@server.command("tcl-lsp.listSubcommands")
-def on_list_subcommands(command_name: str) -> dict:
-    """Return subcommand metadata for a command from the registry."""
-    name = (command_name or "").strip()
-    spec = REGISTRY.get_any(name)
-    if spec is None or not spec.subcommands:
-        return {"command": name, "subcommands": []}
-    return {
-        "command": name,
-        "subcommands": [
-            {
-                "name": sub.name,
-                "detail": sub.detail or "",
-                "synopsis": sub.synopsis or "",
-                "pure": sub.pure,
-                "mutator": sub.mutator,
-                "deprecated": sub.deprecated_replacement is not None,
-            }
-            for sub in sorted(spec.subcommands.values(), key=lambda s: s.name)
-        ],
-    }
-
-
-@server.command("tcl-lsp.listKnownPackages")
-def on_list_known_packages() -> dict:
-    """Return all package names discovered by PackageResolver."""
-    return {
-        "packages": sorted(package_resolver.all_package_names()),
-    }
-
-
-@server.command("tcl-lsp.suggestPackagesForSymbol")
-def on_suggest_packages_for_symbol(symbol: str) -> dict:
-    """Suggest package names for a symbol/command heuristic."""
-    query = (symbol or "").strip()
-    if not query:
-        return {"symbol": query, "suggestions": []}
-
-    suggestions = rank_package_suggestions(
-        query,
-        package_resolver.all_package_names(),
-        20,
-    )
-    return {
-        "symbol": query,
-        "suggestions": suggestions,
-    }
-
-
-@server.command("tcl-lsp.searchHelp")
-def on_search_help(query: str = "", include_screenshots: bool = False) -> dict:
-    """Search the KCS help database for features and documentation."""
-    try:
-        from core.help.kcs_db import (
-            get_feature,
-            get_screenshot_base64,
-            list_features,
-            search_help,
-        )
-    except Exception:
-        return {"error": "KCS help database not available"}
-
-    if query:
-        results = search_help(query)
-        if not results:
-            # Try exact feature lookup
-            feat = get_feature(query)
-            if feat:
-                results = [feat]
-
-        screenshots: dict = {}
-        if include_screenshots:
-            for r in results:
-                from core.help.kcs_db import list_screenshots_for_feature
-
-                for ss in list_screenshots_for_feature(r.get("file", "")):
-                    if ss["has_image"]:
-                        img = get_screenshot_base64(ss["ref_id"])
-                        if img:
-                            screenshots[ss["ref_id"]] = img
-
-        return {"results": results, "screenshots": screenshots}
-
-    catalogue = list_features()
-    return {"catalogue": catalogue}
-
-
-@server.command("tcl-lsp.compilerExplorer")
-def on_compiler_explorer(source: str, dialect: str) -> dict | None:
-    """Run the full compiler explorer pipeline and return serialised JSON."""
-    if not source or not source.strip():
-        return {
-            "error": "No source was received from the editor.",
-            "details": "Open a Tcl/iRule file in the active editor and try again.",
-        }
-    try:
-        result = explorer_run_pipeline(source, dialect=dialect or None)
-        if not result.snapshots:
-            basic_diags, _, _ = get_basic_diagnostics(
-                source,
-                optimiser_enabled=False,
-                disabled_diagnostics=feature_config.disabled_diagnostics,
-                disabled_optimisations=feature_config.disabled_optimisations,
-                line_length=feature_config.line_length,
-            )
-            error_diags = [
-                diagnostic
-                for diagnostic in basic_diags
-                if diagnostic.severity == types.DiagnosticSeverity.Error
-            ]
-            if error_diags:
-                first = error_diags[0]
-                code = first.code if isinstance(first.code, str) else "E000"
-                line = first.range.start.line + 1
-                col = first.range.start.character + 1
-                details = f"{code} at line {line}, column {col}: {first.message}"
-                diagnostics = []
-                for diagnostic in error_diags[:5]:
-                    diag_code = diagnostic.code if isinstance(diagnostic.code, str) else "E000"
-                    diagnostics.append(
-                        {
-                            "code": diag_code,
-                            "message": diagnostic.message,
-                            "line": diagnostic.range.start.line + 1,
-                            "column": diagnostic.range.start.character + 1,
-                        }
-                    )
-                return {
-                    "error": "Source script issues prevented compiler explorer output.",
-                    "details": details,
-                    "diagnostics": diagnostics,
-                }
-            return {
-                "error": "Compiler explorer could not build IR from the active source.",
-                "details": "No compilable Tcl/iRule commands were found in the current editor.",
-            }
-        return explorer_serialise_result(result)
-    except Exception as exc:
-        import traceback
-
-        return {"error": str(exc), "traceback": traceback.format_exc()}
-
-
-@server.command("tcl-lsp.tkPreview")
-def on_tk_preview(source: str) -> dict | None:
-    """Extract Tk widget tree from source code for GUI preview."""
-    if not source or not source.strip():
-        return None
-    try:
-        from core.tk.extract import extract_tk_layout
-
-        return extract_tk_layout(source)
-    except Exception as exc:
-        import traceback
-
-        return {"error": str(exc), "traceback": traceback.format_exc()}
-
-
-@server.command("tcl-lsp.diagramData")
-def on_diagram_data(source: str) -> dict | None:
-    """Extract structured flow data from an iRule for diagram generation."""
-    if not source or not source.strip():
-        return None
-    try:
-        from core.diagram.extract import extract_diagram_data
-
-        return extract_diagram_data(source)
-    except Exception as exc:
-        import traceback
-
-        return {"error": str(exc), "traceback": traceback.format_exc()}
-
-
-@server.command("tcl-lsp.xcTranslate")
-def on_xc_translate(source: str, output_format: str = "both") -> dict | None:
-    """Translate an iRule to F5 XC configuration."""
-    if not source or not source.strip():
-        return None
-    try:
-        from core.xc.json_api import render_json
-        from core.xc.terraform import render_terraform
-        from core.xc.translator import translate_irule
-
-        configure_signatures(dialect="f5-irules")
-        result = translate_irule(source)
-
-        output: dict = {}
-        if output_format in ("terraform", "both"):
-            output["terraform"] = render_terraform(result)
-        if output_format in ("json", "both"):
-            output["json_api"] = render_json(result)
-
-        output["coverage_pct"] = result.coverage_pct
-        output["translatable_count"] = result.translatable_count
-        output["partial_count"] = result.partial_count
-        output["untranslatable_count"] = result.untranslatable_count
-        output["advisory_count"] = result.advisory_count
-        output["items"] = [
-            {
-                "status": item.status.name.lower(),
-                "kind": item.kind.name.lower(),
-                "command": item.irule_command,
-                "xc_description": item.xc_description,
-                "note": item.note,
-                "diagnostic_code": item.diagnostic_code,
-            }
-            for item in result.items
-        ]
-        return output
-    except Exception as exc:
-        import traceback
-
-        return {"error": str(exc), "traceback": traceback.format_exc()}
-
-
-# BIG-IP rule extraction
-
-
-@server.command("tcl-lsp.extractRule")
-def on_extract_rule(uri: str, offset: int) -> dict | None:
-    """Find the ``ltm rule`` / ``gtm rule`` containing *offset* and return it.
-
-    Returns ``{name, fullPath, body, bodyStartOffset, bodyEndOffset}``
-    or ``None`` if the cursor is not inside a rule block.
-    """
-    from core.bigip.rule_extract import find_rule_at_offset
-
-    try:
-        doc = server.workspace.get_text_document(uri)
-    except Exception:
-        return None
-    rule = find_rule_at_offset(doc.source, offset)
-    if rule is None:
-        return None
-    return {
-        "name": rule.name,
-        "fullPath": rule.full_path,
-        "body": rule.body,
-        "bodyStartOffset": rule.body_start_offset,
-        "bodyEndOffset": rule.body_end_offset,
-        "uri": uri,
-    }
-
-
-@server.command("tcl-lsp.listRules")
-def on_list_rules(uri: str) -> list[dict] | None:
-    """Return all ``ltm rule`` / ``gtm rule`` blocks in the given document."""
-    from core.bigip.rule_extract import find_embedded_rules
-
-    try:
-        doc = server.workspace.get_text_document(uri)
-    except Exception:
-        return None
-    rules = find_embedded_rules(doc.source)
-    return [
-        {
-            "name": r.name,
-            "fullPath": r.full_path,
-            "body": r.body,
-            "bodyStartOffset": r.body_start_offset,
-            "bodyEndOffset": r.body_end_offset,
-            "blockStartLine": r.range.start.line,
-            "uri": uri,
-        }
-        for r in rules
-    ]
-
-
-@server.command("tcl-lsp.extractLinkedObjects")
-def on_extract_linked_objects(
-    uri: str,
-    offset: int,
-    max_depth: int = 4,
-    max_nodes: int = 250,
-    extra_offsets: list | None = None,
-) -> dict | None:
-    """Return a transitive BIG-IP object subgraph around one or more cursors.
-
-    *extra_offsets* is an optional list of ``[uri, offset]`` pairs for
-    additional cursor positions beyond the primary *uri*/*offset*.  Each
-    seed object becomes a root in the resulting graph.
-    """
-    from core.bigip.link_extract import extract_linked_bigip_objects
-
-    # Build the full list of (uri, offset) seed positions.
-    seed_positions: list[tuple[str, int]] = [(uri, offset)]
-    if extra_offsets:
-        for item in extra_offsets:
-            if isinstance(item, (list, tuple)) and len(item) >= 2:
-                seed_positions.append((str(item[0]), int(item[1])))
-
-    # Collect all seed URIs whose sources we need.
-    seed_uris: set[str] = {u for u, _ in seed_positions}
-
-    # Ensure every seed URI has a parsed config.
-    configs = background_scanner.bigip_configs
-    for seed_uri in seed_uris:
-        if seed_uri in configs:
-            continue
-        try:
-            src = _get_doc_source(seed_uri)
-        except Exception:
-            return None
-        parsed = background_scanner.parse_bigip_source(seed_uri, src)
-        if parsed is None:
-            return None
-        configs = background_scanner.bigip_configs
-
-    sources: dict[str, str] = {}
-    for cfg_uri in configs:
-        try:
-            sources[cfg_uri] = _get_doc_source(cfg_uri)
-            continue
-        except Exception:
-            pass
-
-        path = uri_to_path(cfg_uri)
-        if not path:
-            continue
-        try:
-            with open(path, encoding="utf-8", errors="replace") as handle:
-                sources[cfg_uri] = handle.read()
-        except OSError:
-            continue
-
-    # Make sure every seed URI is present in sources.
-    for seed_uri in seed_uris:
-        if seed_uri not in sources:
-            try:
-                sources[seed_uri] = _get_doc_source(seed_uri)
-            except Exception:
-                return None
-
-    return extract_linked_bigip_objects(
-        offsets=seed_positions,
-        sources=sources,
-        configs=configs,
-        max_depth=max_depth,
-        max_nodes=max_nodes,
-    )
-
-
-@server.command("tcl-lsp.writeRuleBack")
-def on_write_rule_back(
-    uri: str,
-    body_start_offset: int,
-    body_end_offset: int,
-    new_body: str,
-) -> bool:
-    """Replace the iRule body in the configuration file.
-
-    Applies a ``workspace/applyEdit`` to replace the text between
-    *body_start_offset* and *body_end_offset* with *new_body*.
-    """
-    try:
-        doc = server.workspace.get_text_document(uri)
-        source = doc.source
-    except Exception:
-        return False
-
-    buf = DocumentBuffer.from_source(source)
-    start_pos = buf.offset_to_position(body_start_offset)
-    end_pos = buf.offset_to_position(body_end_offset)
-    start = types.Position(line=start_pos.line, character=start_pos.character)
-    end = types.Position(line=end_pos.line, character=end_pos.character)
-
-    edit = types.WorkspaceEdit(
-        changes={
-            uri: [
-                types.TextEdit(
-                    range=types.Range(start=start, end=end),
-                    new_text=new_body,
-                )
-            ]
-        }
-    )
-    server.workspace_apply_edit(
-        types.ApplyWorkspaceEditParams(edit=edit, label="Write iRule back to config")
-    )
-    return True
-
-
-# tclpkg package manager commands
-
-
-@server.command("tcl-lsp.tclpkg.install")
-def on_tclpkg_install(package_name: str, uri: str = "") -> dict:
-    """Install a Tcl package via the tclpkg manifest.
-
-    This is a stub handler that the VS Code extension or other clients
-    invoke via the ``Install via tclpkg`` code action.  The full
-    implementation will call into ``tclpkg.manifest`` and
-    ``tclpkg.resolver`` in-process; for now it returns a status payload.
-    """
-    return {
-        "success": False,
-        "message": f"tclpkg install for '{package_name}' not yet wired (use the CLI: tcl pkg add {package_name})",
-    }
-
-
-@server.command("tcl-lsp.tclpkg.search")
-def on_tclpkg_search(query: str) -> dict:
-    """Search the tclpkg registry for packages matching *query*."""
-    try:
-        from core.common.user_config import _cache_dir
-        from tclpkg.registry import RegistryClient
-
-        client = RegistryClient(_cache_dir(), offline=True)
-        results = client.search(query)
-        return {"results": [{"name": e.name, "description": e.description} for e in results[:20]]}
-    except Exception as exc:
-        return {"results": [], "error": str(exc)}
-
-
 # Formatting
 
 
@@ -1864,784 +1159,6 @@ def on_will_save_wait_until(
         lines=state.lines if state else None,
     )
     return edits or None
-
-
-
-@server.feature(types.TEXT_DOCUMENT_DID_OPEN)
-async def did_open(params: types.DidOpenTextDocumentParams) -> None:
-    t_open = time.perf_counter()
-    uri = params.text_document.uri
-    lang_id = params.text_document.language_id or ""
-    n_lines = params.text_document.text.count("\n") + 1
-    log.info("Opened %s (language_id=%r, lines=%d)", uri, lang_id, n_lines)
-    # Pre-create the document state so the language_id is stored before
-    # _publish_diagnostics calls workspace_state.update (which would
-    # otherwise create the entry without one).
-    #
-    # analyse=False keeps the event loop free — the full analysis runs
-    # later in _publish_diagnostics via asyncio.to_thread.
-    workspace_state.open(
-        uri,
-        params.text_document.text,
-        params.text_document.version,
-        language_id=lang_id,
-        analyse=False,
-    )
-
-    # Auto-detect dialect from unambiguous file extension (always) or
-    # from source content (only when the user hasn't explicitly configured
-    # a dialect).  The .irul extension is a strong enough signal to
-    # override the editor's default setting — the pull-model config
-    # reports the package.json default ("tcl8.6") which would otherwise
-    # suppress auto-detection via dialect_explicitly_set.
-    if not is_irules_dialect() and _dp._is_irules_source(uri):
-        log.info("Auto-switching to f5-irules dialect (language_id=%r)", lang_id)
-        configure_signatures(dialect="f5-irules")
-        server.window_show_message(
-            types.ShowMessageParams(
-                type=types.MessageType.Info,
-                message="Switched to iRules dialect for F5 iRules support.",
-            )
-        )
-    elif not feature_config.dialect_explicitly_set:
-        from core.common.dialect import detect_dialect_from_source
-
-        source_dialect = detect_dialect_from_source(params.text_document.text)
-        if source_dialect:
-            changed = configure_signatures(dialect=source_dialect)
-            if changed:
-                log.info(
-                    "Auto-switched to %s dialect (detected from source)",
-                    source_dialect,
-                )
-        else:
-            # No dialect hint in this file — check workspace for a higher
-            # Tcl version (e.g. another file has package require Tcl 9.0).
-            _upgrade_dialect_from_workspace()
-
-    if _dp._is_bigip_conf(uri):
-        _dp._publish_bigip_diagnostics(
-            uri,
-            params.text_document.text,
-            params.text_document.version,
-        )
-        return
-    if _dp._is_apl_source(uri):
-        _dp._publish_apl_diagnostics(
-            uri,
-            params.text_document.text,
-            params.text_document.version,
-        )
-        return
-    # Fire-and-forget: let the event loop serve requests immediately
-    # (semantic tokens, document symbols, etc.) while analysis runs in
-    # a background thread.  Awaiting here would block this coroutine
-    # for the entire analysis duration (~seconds on large files),
-    # starving the event loop due to GIL contention.
-    asyncio.create_task(
-        _dp._publish_diagnostics(
-            uri,
-            params.text_document.text,
-            params.text_document.version,
-        )
-    )
-    log.info("[timing] did_open total %.0fms (uri=%s)", (time.perf_counter() - t_open) * 1000, uri)
-
-
-@server.feature(types.TEXT_DOCUMENT_DID_CHANGE)
-async def did_change(params: types.DidChangeTextDocumentParams) -> None:
-    doc = server.workspace.get_text_document(params.text_document.uri)
-    if _dp._is_bigip_conf(params.text_document.uri):
-        _dp._publish_bigip_diagnostics(
-            params.text_document.uri,
-            doc.source,
-            params.text_document.version,
-        )
-        return
-    if _dp._is_apl_source(params.text_document.uri):
-        _dp._publish_apl_diagnostics(
-            params.text_document.uri,
-            doc.source,
-            params.text_document.version,
-        )
-        return
-    await _dp._publish_diagnostics(
-        params.text_document.uri,
-        doc.source,
-        params.text_document.version,
-    )
-
-
-@server.feature(types.TEXT_DOCUMENT_DID_CLOSE)
-def did_close(params: types.DidCloseTextDocumentParams) -> None:
-    uri = params.text_document.uri
-    log.info("Closed %s", uri)
-    diagnostic_scheduler.cancel(uri)
-    if _dp._is_bigip_conf(uri):
-        background_scanner.remove_bigip_config(uri)
-        server.text_document_dp._publish_diagnostics(
-            types.PublishDiagnosticsParams(uri=uri, diagnostics=[])
-        )
-        return
-    if _dp._is_apl_source(uri):
-        background_scanner.remove_apl_model(uri)
-        server.text_document_dp._publish_diagnostics(
-            types.PublishDiagnosticsParams(uri=uri, diagnostics=[])
-        )
-        return
-    with _semantic_token_results_lock:
-        _semantic_token_results.pop(uri, None)
-    workspace_state.close(uri)
-    # If this file was also scanned in the background, revert to that entry
-    # so cross-file references keep working after the file is closed.
-    bg_analysis = background_scanner.get_cached(uri)
-    if bg_analysis is not None:
-        workspace_index.update(uri, bg_analysis, EntrySource.BACKGROUND)
-    else:
-        workspace_index.remove(uri)
-    # Clear diagnostics for closed file
-    server.text_document_dp._publish_diagnostics(
-        types.PublishDiagnosticsParams(uri=uri, diagnostics=[])
-    )
-
-
-# Workspace scanning
-
-
-def _warm_imports() -> None:
-    """Eagerly import heavy modules to eliminate cold-start latency.
-
-    The ``lsp.features.diagnostics`` module transitively imports
-    ``lsprotocol.types`` which compiles hundreds of attrs classes on
-    first load (~500ms).  By triggering these imports during server
-    initialisation (in a daemon thread, concurrent with workspace
-    scanning), the first ``didOpen``/``didChange`` avoids paying this
-    cost on the hot path.
-    """
-    t0 = time.perf_counter()
-    try:
-        # These imports trigger the heavy attrs class compilation in
-        # lsprotocol.types and the diagnostic/compiler pipelines.
-        import core.analysis  # noqa: F811, F401
-        import core.compiler.compilation_unit  # noqa: F811, F401
-        import lsp.features._semantic_tokens  # noqa: F811, F401
-        import lsp.features.diagnostics  # noqa: F811, F401
-
-        elapsed_ms = (time.perf_counter() - t0) * 1000
-        log.info("[timing] import warming %.0fms", elapsed_ms)
-    except Exception:
-        log.debug("Import warming failed (non-fatal)", exc_info=True)
-
-
-_scan_lock = threading.Lock()
-
-_TCL_VERSION_ORDER = {"tcl8.4": 0, "tcl8.5": 1, "tcl8.6": 2, "tcl9.0": 3}
-
-
-def _upgrade_dialect_from_workspace() -> None:
-    """Upgrade the global dialect if any workspace file requires a higher Tcl version.
-
-    Scans ``package_requires`` across all indexed files for ``Tcl``
-    version requirements and switches the dialect when a higher version
-    is found (e.g. ``package require Tcl 9.0`` upgrades from tcl8.6 to
-    tcl9.0).
-    """
-    from core.common.dialect import active_dialect
-
-    current = active_dialect()
-    current_rank = _TCL_VERSION_ORDER.get(current, 2)
-    best_dialect = current
-    best_rank = current_rank
-
-    for uri in workspace_index.all_uris():
-        analysis = workspace_index.get_analysis(uri)
-        if analysis is None:
-            continue
-        for pr in analysis.package_requires:
-            if pr.name != "Tcl":
-                continue
-            ver = pr.version or ""
-            # Strip trailing range markers (e.g. "9.0-" → "9.0")
-            ver_clean = ver.rstrip("-")
-            dialect_key = f"tcl{ver_clean}" if not ver_clean.startswith("tcl") else ver_clean
-            rank = _TCL_VERSION_ORDER.get(dialect_key)
-            if rank is not None and rank > best_rank:
-                best_rank = rank
-                best_dialect = dialect_key
-
-    if best_dialect != current:
-        changed = configure_signatures(dialect=best_dialect)
-        if changed:
-            log.info(
-                "Auto-upgraded dialect to %s (detected from workspace Tcl version)",
-                best_dialect,
-            )
-
-
-def _run_background_scan(
-    progress_cb: Callable[[int, int, str], None] | None = None,
-) -> None:
-    """Execute background scan in a daemon thread.
-
-    Each file is analysed with a per-file timeout so that a single
-    pathological file cannot block the entire scan.  The daemon thread
-    runs independently of the asyncio event loop — the GIL switch
-    interval ensures the event loop stays responsive.
-
-    Concurrent scan requests are deduplicated: if a scan is already
-    running, subsequent calls return immediately.
-
-    ``progress_cb`` is an optional callback invoked with ``(idx, total,
-    path)`` after each file is processed.  When supplied it should be
-    thread-safe; the canonical use is to marshal a
-    ``$/progress`` notification onto the event loop.
-    """
-    if not _scan_lock.acquire(blocking=False):
-        log.debug("Background scan already running — skipping duplicate request")
-        return
-    t0 = time.perf_counter()
-    try:
-        open_uris = frozenset(uri for uri, _ in workspace_state.items())
-        results = background_scanner.scan_all(skip_uris=open_uris, progress_cb=progress_cb)
-        for uri, scan_result in results.items():
-            # Don't overwrite entries for currently open files
-            if workspace_state.get(uri) is not None:
-                continue
-            workspace_index.update(uri, scan_result.analysis, EntrySource.BACKGROUND)
-        # Register iRules procs as globally available
-        for uri, procs in background_scanner.irules_procs.items():
-            workspace_index.update_irules_globals(uri, procs)
-        # Register RULE_INIT variables as cross-file available
-        for uri, exports in background_scanner.irules_rule_init_vars.items():
-            workspace_index.update_rule_init_vars(uri, exports)
-        # Index files discovered via tclIndex that were not already scanned.
-        auto_loaded = 0
-        for _proc_name, source_file in background_scanner.auto_index_entries.items():
-            source_uri = path_to_uri(source_file)
-            if workspace_index.get_analysis(source_uri) is not None:
-                continue
-            scan_result = background_scanner.rescan_file(source_file)
-            if scan_result:
-                workspace_index.update(
-                    source_uri,
-                    scan_result.analysis,
-                    EntrySource.AUTO_INDEX,
-                )
-                auto_loaded += 1
-        # Auto-upgrade dialect based on workspace-level Tcl version.
-        if not feature_config.dialect_explicitly_set:
-            _upgrade_dialect_from_workspace()
-        elapsed_ms = (time.perf_counter() - t0) * 1000
-        log.info(
-            "[timing] background scan %.0fms (%d files, %d auto-index)",
-            elapsed_ms,
-            len(results),
-            auto_loaded,
-        )
-    except Exception:
-        log.error("Background scan failed", exc_info=True)
-    finally:
-        _scan_lock.release()
-
-
-# Strong reference for the scan coordinator task so asyncio does not
-# garbage-collect it mid-run.  Reset on each new scan; we only ever have
-# one coordinator alive at a time because ``_run_background_scan`` uses
-# ``_scan_lock`` to deduplicate concurrent calls.
-_scan_progress_task: "asyncio.Task[None] | None" = None
-
-
-def _start_scan_with_progress() -> None:
-    """Run the background scan with $/progress notifications.
-
-    The scan itself runs in a daemon thread (same as the silent path).  A
-    tiny coordinator runs on the event loop: it issues
-    ``window/workDoneProgress/create``, calls ``begin``, kicks off the
-    scan thread, receives progress ticks via ``call_soon_threadsafe``, and
-    emits ``end`` when the thread finishes.
-
-    The coordinator is designed to be *best-effort*: any failure in the
-    ``$/progress`` dance falls through to the plain daemon-thread scan so
-    a misbehaving client never blocks indexing.
-    """
-    import uuid
-
-    global _scan_progress_task
-
-    loop = getattr(server, "loop", None)
-    if loop is None or not loop.is_running():
-        # Fallback: no event loop — run silently.
-        threading.Thread(target=_run_background_scan, daemon=True).start()
-        return
-
-    token = f"tcl-lsp-scan-{uuid.uuid4()}"
-    # True once ``begin`` has successfully been sent to the client.
-    # Reports arriving before this flag is set are dropped rather than
-    # risking an "unknown token" error from the client.
-    progress_active = False
-
-    def _progress_cb(idx: int, total: int, path: str) -> None:
-        if not progress_active:
-            return
-        percentage = int((idx / total) * 100) if total else 0
-
-        def _send() -> None:
-            try:
-                server.work_done_progress.report(
-                    token,
-                    types.WorkDoneProgressReport(
-                        message=f"{idx}/{total}",
-                        percentage=percentage,
-                    ),
-                )
-            except Exception:  # pragma: no cover - defensive
-                pass
-
-        try:
-            loop.call_soon_threadsafe(_send)
-        except Exception:  # pragma: no cover - event loop may be closing
-            pass
-
-    async def _coordinator() -> None:
-        nonlocal progress_active
-        # Give the create_async call a short window to reach the client.
-        # Some clients ignore the request silently; without a timeout the
-        # coordinator (and therefore the scan) would hang forever.
-        try:
-            await asyncio.wait_for(server.work_done_progress.create_async(token), timeout=5.0)
-        except (Exception, asyncio.TimeoutError):
-            log.debug(
-                "work_done_progress/create failed or timed out; running scan silently",
-                exc_info=True,
-            )
-            threading.Thread(target=_run_background_scan, daemon=True).start()
-            return
-
-        try:
-            server.work_done_progress.begin(
-                token,
-                types.WorkDoneProgressBegin(
-                    title="Scanning workspace",
-                    cancellable=False,
-                    percentage=0,
-                ),
-            )
-            progress_active = True
-        except Exception:  # pragma: no cover - defensive
-            # ``begin`` failed — bail out to the silent path.
-            threading.Thread(target=_run_background_scan, daemon=True).start()
-            return
-
-        done = threading.Event()
-
-        def _worker() -> None:
-            try:
-                _run_background_scan(progress_cb=_progress_cb)
-            finally:
-                done.set()
-
-        threading.Thread(target=_worker, daemon=True).start()
-
-        # Wait for completion without blocking the event loop.
-        while not done.is_set():
-            await asyncio.sleep(0.1)
-
-        progress_active = False
-        try:
-            server.work_done_progress.end(
-                token,
-                types.WorkDoneProgressEnd(message="done"),
-            )
-        except Exception:  # pragma: no cover - defensive
-            pass
-
-    # Keep a strong reference so the task does not get garbage-collected
-    # if the enclosing scope is dropped before the coroutine finishes.
-    # ``loop.create_task`` avoids the deprecated ``loop=`` keyword on
-    # ``asyncio.ensure_future`` and does exactly what we need: schedule
-    # the coroutine on the already-running loop.
-    _scan_progress_task = loop.create_task(_coordinator())
-
-
-@server.feature(types.INITIALIZED)
-def on_initialized(params: types.InitializedParams) -> None:
-    """After client initialization, scan workspace for Tcl files."""
-    from core.common.dialect import active_dialect
-
-    # Load user config from the platform-native config file.
-    user_config = load_user_config()
-
-    # Apply config-file settings as baseline defaults.  Editor settings
-    # received later via ``didChangeConfiguration`` will override these.
-    config_settings = get_all_settings(user_config)
-    if config_settings:
-        formatting = config_settings.get("formatting")
-        if isinstance(formatting, dict) and formatting:
-            _state.formatter_config = FormatterConfig.from_dict(_normalise_formatter_settings(formatting))
-        _apply_feature_settings(config_settings)
-
-    process_start = getattr(server, "_process_start_time", None)
-    if process_start is not None:
-        startup_ms = (time.monotonic() - process_start) * 1000
-        log.info("[timing] server ready: %.0fms from process start", startup_ms)
-
-    log.info(
-        "Server initialized (version=%s, dialect=%s)",
-        _version,
-        active_dialect(),
-    )
-
-    # Advise editors that don't request semantic tokens.
-    caps = server.client_capabilities
-    st = getattr(getattr(caps, "text_document", None), "semantic_tokens", None)
-    if st is None:
-        log.info("Client did not advertise semantic token support")
-        server.window_show_message(
-            types.ShowMessageParams(
-                type=types.MessageType.Info,
-                message=(
-                    "Tip: enable semantic tokens for richer Tcl highlighting. "
-                    'In Zed, add \'"semantic_tokens": "full"\' to your '
-                    "language settings."
-                ),
-            )
-        )
-
-    roots: list[str] = []
-    ws = server.workspace
-    if ws.root_path:
-        roots.append(ws.root_path)
-
-    # tclpkg project detection — look for tclpkg.tcl as a project marker
-    # and auto-add venv/project lib/ directories to library paths.
-    library_paths: list[str] = []
-    if ws.root_path:
-        import os
-
-        # Walk up from root_path to find tclpkg.tcl (monorepo support).
-        _tclpkg_dir = None
-        _search = ws.root_path
-        for _ in range(10):
-            if os.path.isfile(os.path.join(_search, "tclpkg.tcl")):
-                _tclpkg_dir = _search
-                break
-            _parent = os.path.dirname(_search)
-            if _parent == _search:
-                break
-            _search = _parent
-
-        if _tclpkg_dir is not None:
-            log.info("Detected tclpkg project at: %s", _tclpkg_dir)
-            # Add project-local lib/ if present.
-            project_lib = os.path.join(_tclpkg_dir, "lib")
-            if os.path.isdir(project_lib):
-                library_paths.append(project_lib)
-        # Detect active venv: $TCL_VENV or .venv/tclvenv.cfg next to project.
-        venv_dir = os.environ.get("TCL_VENV")
-        if not venv_dir:
-            local_venv = os.path.join(ws.root_path, ".venv", "tclvenv.cfg")
-            if os.path.isfile(local_venv):
-                venv_dir = os.path.join(ws.root_path, ".venv")
-        if venv_dir:
-            venv_lib = os.path.join(venv_dir, "lib")
-            if os.path.isdir(venv_lib):
-                library_paths.append(venv_lib)
-                log.info("Auto-detected tclpkg venv lib: %s", venv_lib)
-
-    background_scanner.configure(
-        workspace_roots=roots,
-        library_paths=library_paths if library_paths else None,
-    )
-    # Prepend workspace roots to the package resolver's search paths so
-    # a workspace-local copy of a package (e.g. someone editing tcllib)
-    # wins over any copy provided by the configured library paths —
-    # matching Tcl's own ``auto_path`` semantics where the first
-    # provider found satisfies ``package require``.
-    resolver_paths = list(roots) + list(library_paths)
-    if resolver_paths:
-        package_resolver.configure(search_paths=resolver_paths)
-        _loaded_packages.clear()
-
-    # Kick off the scan with optional progress reporting.  When the client
-    # advertises window/workDoneProgress we emit begin/report/end notifications
-    # under a fresh token.  Otherwise we just run the silent daemon path as
-    # before.
-    progress_supported = bool(
-        feature_config.progress_enabled
-        and getattr(getattr(caps, "window", None), "work_done_progress", False)
-    )
-    if progress_supported:
-        _start_scan_with_progress()
-    else:
-        threading.Thread(target=_run_background_scan, daemon=True).start()
-
-    # Eagerly import heavy modules in a background thread so that the
-    # first didOpen/didChange does not pay the ~500ms import cost for
-    # lsprotocol.types, attrs class compilation, and the diagnostics
-    # module.  This runs concurrently with workspace scanning.
-    threading.Thread(target=_warm_imports, daemon=True).start()
-
-    # Proactively pull editor settings so that clients using the pull model
-    # (vscode-languageclient v10, JetBrains) have their settings applied
-    # immediately rather than waiting for a configuration change event.
-    _pull_and_apply_configuration()
-
-
-# Custom commands (workspace/executeCommand)
-
-_DIALECT_COMMAND = "tcl-lsp.setDialect"
-_EXPORT_CONFIG_COMMAND = "tcl-lsp.exportConfig"
-_DIALECT_LABELS = {
-    "tcl8.4": "Tcl 8.4",
-    "tcl8.5": "Tcl 8.5",
-    "tcl8.6": "Tcl 8.6",
-    "tcl9.0": "Tcl 9.0",
-    "f5-irules": "F5 iRules",
-    "f5-iapps": "F5 iApps",
-    "eda-tools": "EDA Tools",
-}
-
-
-@server.command(_DIALECT_COMMAND)
-def on_set_dialect(dialect: str = "") -> dict:
-    """Switch the active Tcl dialect."""
-    return _switch_dialect(dialect)
-
-
-@server.command(_EXPORT_CONFIG_COMMAND)
-def on_export_config() -> dict:
-    """Export the current server configuration."""
-    return _export_config()
-
-
-def _switch_dialect(dialect: str) -> dict:
-    """Switch the active dialect and re-publish diagnostics."""
-    from core.commands.registry.dialects import KNOWN_DIALECTS
-    from core.common.dialect import active_dialect
-
-    if dialect and dialect not in KNOWN_DIALECTS:
-        return {"success": False, "error": f"Unknown dialect: {dialect!r}"}
-
-    prev = active_dialect()
-    changed = configure_signatures(dialect=dialect or None)
-    current = active_dialect()
-    feature_config.dialect_explicitly_set = True
-    log.info("Dialect set to %s (was %s)", current, prev)
-
-    if changed:
-        server.window_show_message(
-            types.ShowMessageParams(
-                type=types.MessageType.Info,
-                message=f"Switched dialect to {_DIALECT_LABELS.get(current, current)}.",
-            )
-        )
-        # Re-analyse all open documents with the new dialect.
-        # Use async tasks so the event loop stays responsive.
-        diagnostic_scheduler.cancel_all()
-        try:
-            loop = asyncio.get_running_loop()
-        except RuntimeError:
-            loop = None
-        for uri, state in workspace_state.items():
-            if loop is not None:
-                loop.create_task(
-                    _dp._publish_diagnostics(
-                        uri,
-                        state.source,
-                        state.version,
-                        force_reanalyse=True,
-                    )
-                )
-            else:
-                _dp._publish_diagnostics_sync(
-                    uri,
-                    state.source,
-                    state.version,
-                    force_reanalyse=True,
-                )
-
-    return {"success": True, "dialect": current}
-
-
-def _export_config() -> dict:
-    """Export the current effective settings to the config file."""
-
-    # Build the current effective settings dict.
-    settings: dict[str, object] = {}
-
-    # Diagnostics — collect disabled codes.
-    diag: dict[str, object] = {}
-    for code in _ALL_DIAGNOSTIC_CODES:
-        if code in feature_config.disabled_diagnostics:
-            diag[code] = False
-    current_patterns = list(feature_config.generic_variable_patterns)
-    if current_patterns and current_patterns != list(DEFAULT_GENERIC_VARIABLE_PATTERNS):
-        diag["genericVariablePatterns"] = current_patterns
-    if diag:
-        settings["diagnostics"] = diag
-
-    # Optimiser — export profile + only per-code overrides relative to profile.
-    opt: dict[str, object] = {"enabled": feature_config.optimiser_enabled}
-    opt["profile"] = feature_config.optimiser_profile
-    profile_baseline = profile_to_disabled(profile_from_name(feature_config.optimiser_profile))
-    for code in _ALL_OPTIMISATION_CODES:
-        in_disabled = code in feature_config.disabled_optimisations
-        in_baseline = code in profile_baseline
-        if in_disabled and not in_baseline:
-            opt[code] = False  # explicitly disabled beyond profile
-        elif not in_disabled and in_baseline:
-            opt[code] = True  # explicitly enabled beyond profile
-    settings["optimiser"] = opt
-
-    # Shimmer
-    settings["shimmer"] = {"enabled": feature_config.shimmer_enabled}
-
-    # XC Diagnostics
-    settings["xcDiagnostics"] = {"enabled": feature_config.xc_diagnostics_enabled}
-
-    # Features
-    feat: dict[str, object] = {}
-    for json_key, attr in _FEATURE_TOGGLE_KEYS.items():
-        feat[json_key] = getattr(feature_config, attr)
-    settings["features"] = feat
-
-    # Style
-    settings["style"] = {"lineLength": feature_config.line_length}
-
-    # Build defaults so only non-default values are written.
-    default_cfg = FeatureConfig()
-    default_features: dict[str, object] = {
-        json_key: getattr(default_cfg, attr) for json_key, attr in _FEATURE_TOGGLE_KEYS.items()
-    }
-    defaults: dict[str, object] = {
-        "features": default_features,
-        "optimiser": {
-            "enabled": default_cfg.optimiser_enabled,
-            "profile": default_cfg.optimiser_profile,
-        },
-        "shimmer": {"enabled": default_cfg.shimmer_enabled},
-        "xcDiagnostics": {"enabled": default_cfg.xc_diagnostics_enabled},
-        "style": {"lineLength": default_cfg.line_length},
-    }
-
-    try:
-        path = save_settings_to_config(settings, only_non_default=True, defaults=defaults)
-        server.window_show_message(
-            types.ShowMessageParams(
-                type=types.MessageType.Info,
-                message=f"Settings exported to {path}",
-            )
-        )
-        return {"success": True, "path": path}
-    except Exception as exc:
-        log.error("Failed to export config", exc_info=True)
-        return {"success": False, "error": str(exc)}
-
-
-@server.feature(types.SHUTDOWN)
-def on_shutdown(params: None) -> None:
-    """Cancel pending background tasks and shut down process pool."""
-    diagnostic_scheduler.cancel_all()
-    global _process_pool
-    if _process_pool is not None:
-        _process_pool.shutdown(wait=False, cancel_futures=True)
-        _process_pool = None
-
-
-@server.feature(types.WORKSPACE_DID_CHANGE_WATCHED_FILES)
-def did_change_watched_files(
-    params: types.DidChangeWatchedFilesParams,
-) -> None:
-    """React to file system changes for non-open files."""
-    for change in params.changes:
-        uri = change.uri
-        # Skip files that are currently open (they get fresh analysis on edit)
-        if workspace_state.get(uri) is not None:
-            continue
-
-        if change.type == types.FileChangeType.Deleted:
-            workspace_index.remove(uri)
-            background_scanner.remove_file(uri)
-        elif change.type in (
-            types.FileChangeType.Created,
-            types.FileChangeType.Changed,
-        ):
-            file_path = uri_to_path(uri)
-            if file_path:
-                scan_result = background_scanner.rescan_file(file_path)
-                if scan_result:
-                    workspace_index.update(
-                        uri,
-                        scan_result.analysis,
-                        EntrySource.BACKGROUND,
-                    )
-                    # Update iRules globals if applicable
-                    if scan_result.dialect_hint == "f5-irules":
-                        workspace_index.update_irules_globals(
-                            uri,
-                            scan_result.analysis.all_procs,
-                        )
-                        if scan_result.rule_init_exports:
-                            workspace_index.update_rule_init_vars(
-                                uri,
-                                scan_result.rule_init_exports,
-                            )
-
-
-# File rename hooks
-
-
-_RENAME_FILE_OPERATION_OPTIONS = types.FileOperationRegistrationOptions(
-    filters=[
-        types.FileOperationFilter(
-            pattern=types.FileOperationPattern(
-                glob="**/*.{tcl,tm,itcl,irule,irul}",
-            ),
-        ),
-    ],
-)
-
-
-@server.feature(types.WORKSPACE_WILL_RENAME_FILES, _RENAME_FILE_OPERATION_OPTIONS)
-def on_will_rename_files(
-    params: types.RenameFilesParams,
-) -> types.WorkspaceEdit | None:
-    """Rewrite ``source`` lines in dependents so the workspace still loads."""
-    if not feature_config.workspace_file_ops_enabled:
-        return None
-    roots: list[str] = []
-    ws = server.workspace
-    if ws.root_path:
-        roots.append(ws.root_path)
-    return compute_batch_rename_edits(
-        list(params.files),
-        workspace_index,
-        workspace_roots=roots,
-    )
-
-
-@server.feature(types.WORKSPACE_DID_RENAME_FILES, _RENAME_FILE_OPERATION_OPTIONS)
-def on_did_rename_files(params: types.RenameFilesParams) -> None:
-    """Reindex renamed files after the client applies the rename on disk."""
-    if not feature_config.workspace_file_ops_enabled:
-        return
-    for f in params.files:
-        old_uri = f.old_uri
-        new_uri = f.new_uri
-        workspace_index.remove(old_uri)
-        background_scanner.remove_file(old_uri)
-        new_path = uri_to_path(new_uri)
-        if not new_path:
-            continue
-        scan_result = background_scanner.rescan_file(new_path)
-        if scan_result:
-            workspace_index.update(
-                new_uri,
-                scan_result.analysis,
-                EntrySource.BACKGROUND,
-            )
-
-
 
 # Configuration (extracted to lsp/settings.py)
